Big Decimal

+++ src/main/kotlin/br/com/dastec/gerenciasalao/controllers/requests/payments/PostPaymentServiceRequest.kt
@@ -2,6 +2,7 @@ package br.com.dastec.gerenciasalao.controllers.requests.payments
 
 import br.com.dastec.gerenciasalao.validation.annotation.IsCurrency
 import com.fasterxml.jackson.annotation.JsonAlias
+import java.math.BigDecimal
 import javax.validation.constraints.NotNull
 
 
@@ -17,5 +18,5 @@ data class PostPaymentServiceRequest(
 
     @JsonAlias("value_payment")
     @IsCurrency
-    var valuePayment: Double,
+    var valuePayment: BigDecimal,
 )
