+++ src/main/kotlin/br/com/dastec/gerenciasalao/controllers/requests/saleServices/PutUpdateSaleServiceRequest.kt
@@ -1,9 +1,10 @@
 package br.com.dastec.gerenciasalao.controllers.requests
 
 import br.com.dastec.gerenciasalao.validation.annotation.IsCurrency
+import java.math.BigDecimal
 
 
 data class PutUpdateSaleServiceRequest(
     @IsCurrency
-    val price: Double
+    val price: BigDecimal
 )
